+++ src/main/java/cotato/bookitlist/post/dto/requeset/PostUpdateRequest.java
@@ -1,4 +1,4 @@
-package cotato.bookitlist.post.dto;
+package cotato.bookitlist.post.dto.requeset;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
